% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWeight.R
\name{calcWeight}
\alias{calcWeight}
\title{Calculate weights for a set of internal controls}
\usage{
calcWeight(x)
}
\arguments{
\item{x}{A matrix with each row corresponding to an internal control}
}
\value{
a vector containing the weights of each internal control
}
\description{
Calculate weights for a set of internal controls
}
\examples{
x <- matrix(1:6, 2, 3)
w <- calcWeight(x)
}
