% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GSE50013}
\alias{data_GSE50013}
\title{GSE50013 Liver Cancer qPCR array (Plasma)}
\format{
## `data_GSE50013`
A matrix with 241 rows and 40 columns:
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE50013>
}
\usage{
data_GSE50013
}
\description{
A subset of GSE50013
miRs with ct values more than 39 in at least one sample and
miRs with ct values more than 35 in all samples are removed.
miRs with at most 3 missing values were imputed using nonedetects package
}
\keyword{datasets}
