% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GSE57661}
\alias{data_GSE57661}
\title{GSE57661 Early-stage breast cancer qPCR array (Plasma)}
\format{## `data_GSE57661`
A matrix with 176 rows and 48 columns:}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57661>
}
\usage{
data_GSE57661
}
\description{
A subset of GSE57661
miRs with ct values more than 39 in at least one sample and
miRs with ct values more than 35 in all samples are removed.
miRs with at most 3 missing values were imputed using nonedetects package
}
\keyword{datasets}
