% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GSE67075}
\alias{data_GSE67075}
\title{GSE67075 Colon Cancer qPCR array (Plasma)}
\format{
## `data_GSE67075`
A matrix with 554 rows and 48 columns:
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE67075>
}
\usage{
data_GSE67075
}
\description{
A subset of GSE67075
miRs with ct values more than 39 in at least one sample and
miRs with ct values more than 35 in all samples are removed.
miRs with at most 3 missing values were imputed using nonedetects package
}
\keyword{datasets}
