% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GSE78870}
\alias{data_GSE78870}
\title{GSE78870 Primary breast cancer qPCR array (Tissue)}
\format{
## `data_GSE78870`
A matrix with 364 rows and 106 columns:
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE78870>
}
\usage{
data_GSE78870
}
\description{
A subset of GSE78870
miRs with ct values more than 39 in at least one sample and
miRs with ct values more than 35 in all samples are removed.
miRs with at most 3 missing values were imputed using nonedetects package
}
\keyword{datasets}
