% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GSE90828}
\alias{data_GSE90828}
\title{GSE90828 Colon Cancer qPCR array (Plasma)}
\format{## `data_GSE90828`
A matrix with 336 rows and 53 columns:}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE90828>
}
\usage{
data_GSE90828
}
\description{
A subset of GSE90828
miRs with ct values more than 39 in at least one sample and
miRs with ct values more than 35 in all samples are removed.
miRs with at most 3 missing values were imputed using nonedetects package
}
\keyword{datasets}
