% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_TCGA_BRCA}
\alias{data_TCGA_BRCA}
\title{TCGA Breast Cancer miRNA-Seq data (Tissue)}
\format{## `data_TCGA_BRCA`
A matrix with 185 rows and 1201 columns:}
\usage{
data_TCGA_BRCA
}
\description{
the data is normalized in counts per million (cpm)
}
\keyword{datasets}
