% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups_GSE50013}
\alias{groups_GSE50013}
\title{sample groups of data_GSE50013 data}
\format{
## `groups_GSE50013`
A factor with 40 elements
}
\usage{
groups_GSE50013
}
\description{
each element represents the group of the sample
}
\keyword{datasets}
