% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups_GSE57661}
\alias{groups_GSE57661}
\title{sample groups of data_GSE57661 data}
\format{
## `groups_GSE57661`
A factor with 48 elements
}
\usage{
groups_GSE57661
}
\description{
each element represents the group of the sample
}
\keyword{datasets}
