% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups_GSE59520}
\alias{groups_GSE59520}
\title{sample groups of data_GSE59520 data}
\format{
## `groups_GSE59520`
A factor with 48 elements
}
\usage{
groups_GSE59520
}
\description{
each element represents the group of the sample
}
\keyword{datasets}
