% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups_GSE67075}
\alias{groups_GSE67075}
\title{sample groups of data_GSE67075 data}
\format{## `groups_GSE67075`
A factor with 48 elements}
\usage{
groups_GSE67075
}
\description{
each element represents the group of the sample
}
\keyword{datasets}
