% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups_GSE78870}
\alias{groups_GSE78870}
\title{sample groups of data_GSE78870 data}
\format{## `groups_GSE78870`
A factor with 106 elements}
\usage{
groups_GSE78870
}
\description{
each element represents the group of the sample
}
\keyword{datasets}
