% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups_GSE90828}
\alias{groups_GSE90828}
\title{sample groups of data_GSE90828 data}
\format{## `groups_GSE90828`
A factor with 53 elements}
\usage{
groups_GSE90828
}
\description{
each element represents the group of the sample
}
\keyword{datasets}
