% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups_TCGA_BRCA}
\alias{groups_TCGA_BRCA}
\title{sample groups of data_TCGA_BRCA data}
\format{
## `groups_TCGA_BRCA`
A matrix with 1201 rows and 7 columns:
}
\usage{
groups_TCGA_BRCA
}
\description{
each row represents the tumor subtype of each sample
}
\keyword{datasets}
